API/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
certifi==2020.6.20
wheel
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
GeoAlchemy2==0.8.4
gunicorn==20.0.4
idna==2.10
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.1.2
psycopg2==2.8.4
py-postgresql==1.2.1
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.10
Werkzeug==1.0.1
